+++ modules/admin-ui/src/main/webapp/scripts/modules/events/controllers/eventController.js
@@ -24,13 +24,13 @@
 angular.module('adminNg.controllers')
 .controller('EventCtrl', [
     '$scope', 'Notifications', 'EventTransactionResource', 'EventMetadataResource', 'EventAssetsResource',
-    'EventCatalogsResource', 'CommentResource', 'EventWorkflowsResource', 'EventWorkflowActionResource', 'EventWorkflowDetailsResource',
-    'ResourcesListResource', 'UserRolesResource', 'EventAccessResource', 'EventGeneralResource',
+    'EventAssetCatalogsResource', 'CommentResource', 'EventWorkflowsResource', 'EventWorkflowActionResource', 'EventWorkflowDetailsResource',
+    'ResourcesListResource', 'UserRolesResource', 'EventAccessResource', 'EventPublicationsResource',
     'OptoutsResource', 'EventParticipationResource', 'EventSchedulingResource', 'NewEventProcessingResource',
     'OptoutSingleResource', 'CaptureAgentsResource', 'ConflictCheckResource', 'Language', 'JsHelper', '$sce', '$timeout', 'EventHelperService',
     'UploadAssetOptions', 'EventUploadAssetResource', 'Table', 'SchedulingHelperService',
-    function ($scope, Notifications, EventTransactionResource, EventMetadataResource, EventAssetsResource, EventCatalogsResource, CommentResource,
-        EventWorkflowsResource, EventWorkflowActionResource, EventWorkflowDetailsResource, ResourcesListResource, UserRolesResource, EventAccessResource, EventGeneralResource,
+    function ($scope, Notifications, EventTransactionResource, EventMetadataResource, EventAssetsResource, EventAssetCatalogsResource, CommentResource,
+        EventWorkflowsResource, EventWorkflowActionResource, EventWorkflowDetailsResource, ResourcesListResource, UserRolesResource, EventAccessResource, EventPublicationsResource,
         OptoutsResource, EventParticipationResource, EventSchedulingResource, NewEventProcessingResource,
         OptoutSingleResource, CaptureAgentsResource, ConflictCheckResource, Language, JsHelper, $sce, $timeout, EventHelperService, UploadAssetOptions,
         EventUploadAssetResource, Table, SchedulingHelperService) {
@@ -229,18 +229,18 @@ angular.module('adminNg.controllers')
             },
             fetchChildResources = function (id) {
 
-                var general = EventGeneralResource.get({ id: id }, function () {
-                    angular.forEach(general.publications, function (publication, index) {
+                var publications = EventPublicationsResource.get({ id: id }, function () {
+                    angular.forEach(publications.publications, function (publication, index) {
                         publication.label = publication.name;
                         publication.order = 999 + index;
                         var now = new Date();
-                        if (publication.id == "engage-live" && 
-                        	(now < new Date(general["start-date"]) || now > new Date(general["end-date"])))
+                        if (publication.id == "engage-live" &&
+                        	(now < new Date(publications["start-date"]) || now > new Date(publications["end-date"])))
                         	publication.enabled = false;
                         else publication.enabled = true;
                     });
                     $scope.publicationChannels = ResourcesListResource.get({ resource: 'PUBLICATION.CHANNELS' }, function() {
-                        angular.forEach(general.publications, function (publication) {
+                        angular.forEach(publications.publications, function (publication) {
                             if(angular.isDefined($scope.publicationChannels[publication.id])) {
                                 var record = JSON.parse($scope.publicationChannels[publication.id]);
                                 if (record.label) publication.label = record.label;
@@ -250,10 +250,10 @@ angular.module('adminNg.controllers')
                                 if (record.order) publication.order = record.order;
                             }
                         });
-                        // we postpone setting $scope.general until this point to avoid UI "flickering" due to publications changing
-                        $scope.general = general;
+                        // we postpone setting $scope.publications until this point to avoid UI "flickering" due to publications changing
+                        $scope.publications = publications;
                     }, function() {
-                        $scope.general = general;
+                        $scope.publications = publications;
                     });
                 });