+++ docs/guides/admin/docs/releasenotes.md
@@ -13,6 +13,9 @@ Configuration changes
 - The tracking options defaults have changed to be more aware of the European Union's General Data Protection
   Regulation. Note that you can [still use the old settings if you want to](configuration/logging.and.privacy.md).
 
+- The role ROLE_UI_EVENTS_DETAILS_GENERAL_VIEW for viewing the publications (previously general) tab in the event
+  details modal has been renamed to ROLE_UI_EVENTS_DETAILS_PUBLICATIONS_VIEW for consistency.
+
 
 Release Schedule
 ----------------